\name{compprocchains}
\alias{compprocchains}

\title{Compare EUMETSAT OSISAF products from two different processing chains}

\description{
    Compare OSISAF products from two different processing chains. This
    function takes two directory names as input, it collects filenames
    from these directories, checks that the file is HDF and that it is
    contained within both directories, before reading it. Data read are
    compared by plotting a time series of the mean differences at the
    active graphics device.
}
\usage{
compprocchains(chain1,chain2,from=ISOdate(1996,1,1),to=Sys.time(),area=NULL) 
}

\arguments{
    \item{chain1}{Full path to first directory to read}
    \item{chain2}{Full path to second directory to read}
    \item{from}{ISOdate specification of oldest file to process}
    \item{to}{ISOdate specification of newest file to process}
    \item{area}{Perl regexp to handle directories with several products
    (e.g. sea ice)}
}
\details{
    Two directories are given as input to the function. Within each
    directory all products fulfilling the time and area specification is
    read. For each product a mean value is calculated and the difference
    in the mean values for the same product within each directory is
    generated.
}
\value{
    Returns a \link{data.frame} containing a time identification and
    difference between the two products.
    \item{time}{Time specification as generated by \link{ISOdatetime}}
    \item{data}{Difference between the two processing chains for each time
    specification}
}
\note{
    This function is probably of most interest to the development team
    when performing software updates.
}
\seealso{\link{readosisaf}}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
    # Compare operational and pre-operational processing
    compprocchains("/data/osisaf/output/ice/multi/conc/",
    "/data/osisaf_it/output/ice/multi/conc/",
    from=ISOdate(2006,4,1),to=ISOdate(2006,6,30),area="sh")
}
