
\name{gammaPDFtableplot}
\alias{gammaPDFtableplot}

\title{Multiple Gamma Probability Density Function plots}

\description{
    Nine figures of the three parameter Gamma distribution with various
    values of the shape, scale and location factors are generated. If
    output to \link{postscript} is chosen, files are automatically
    generated within the directory specifued by basepath. 
}
\usage{
gammaPDFtableplot(postscript=FALSE,basepath)
}

\arguments{
    \item{postscript}{if TRUE, files are automatically generated within
    the basepath directory}
    \item{basepath}{where to put postscript files}
}
\value{A character string}

\seealso{\link{gammaPDFplot}}

\author{
    Øystein Godøy \email{o.godoy@met.no}
}

\examples{}

\keyword{}
