\name{ mycolours }

\alias{ mycolours }

\title{
    mycolours
}

\description{
    Creates a colour scale to be used in plotting. The colour scale is
    known as Jet Colours.
}

\usage{
    mycolours(n=256)
}

\arguments{
    \item{}{}
}

\value{
}

\seealso{
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
