
\name{normalPDFest}
\alias{normalPDFest}

\title{Estimate the probability density function for a dataset}

\description{
Using a Normal distribution, the probability density function is estimated
for a given dataset.
}
\usage{
    normalPDFest(dataset)
}

\arguments{
    \item{dataset}{input data as a vector}
}
\value{Returns a \link{list} containing the mean value and standard
deviation of the input dataset and a vector of the corresponding
probabilities.}

\seealso{}

\author{
    Øystein Godøy \email{o.godoy@met.no}
}

\examples{}

\keyword{}
