
\name{normalPDFplot}
\alias{normalPDFplot}

\title{Plotting of a normal distribution}

\description{
A graphical presentation of the Normal distribution is generated on the
graphical device. 
}
\usage{
normalPDFplot(mean, sdev, xpar) 
}

\arguments{
    \item{mean}{Mean value of the Normal dsitribution}
    \item{sdev}{Standard deviation of the normal distribution}
    \item{xpar}{The values from which the probability density should be
    plotted}
}
\value{NULL}

\seealso{
    \link{normalPDFest}, \link{gammaPDF}, \link{gammaPDFtableplot}, \link{GammaDist}
}

\author{
    Øystein Godøy \email{o.godoy@met.no}
}

\examples{}

\keyword{}
